Migrated Player Test Steps from JumperStepDefinitions.cs to PlayerSteps.cs
